moved slider to separate file
